user-api/requirements.txt
uvicorn==0.13.3
fastapi==0.63.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
psycopg2==2.8.6
pymemcache==3.4.0
pytest==6.2.2
python-dotenv==0.15.0
requests==2.25.1
cryptography==3.2.1
fernet==1.0.1